Add CSName as a proper alias. Change UpTime to a calculated property
Uptime now returns the correct time difference every time it is called.
Only as a string however. Next this should probably return the actualy
timespan object to make filtering easier. It will need a custom display
writing to look nice in the console.

CSName is now aliased as computername, that seems neater.